+++ react/SearchResultFlexible.js
@@ -56,6 +56,9 @@ const SearchResultFlexible = ({
   orderBy,
   page,
   facetsLoading,
+  showFacetQuantity = false,
+  showFacetSearch = false,
+  orderFacetsBy,
 }) => {
   //This makes infinite scroll unavailable.
   //Infinite scroll was deprecated and we have
@@ -100,8 +103,18 @@ const SearchResultFlexible = ({
       hiddenFacets,
       pagination,
       mobileLayout,
+      showFacetQuantity,
+      orderFacetsBy,
+      showFacetSearch,
     }),
-    [hiddenFacets, mobileLayout, pagination]
+    [
+      hiddenFacets,
+      mobileLayout,
+      pagination,
+      showFacetQuantity,
+      orderFacetsBy,
+      showFacetSearch,
+    ]
   )
 
   const context = useMemo(