+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Search input for facets.
+- Prop `orderFacetsBy` to `SearchResultFlexible`.
+
+### Fixed
+- `showFacetQuantity` was not being passed on.
+
 ## [3.39.10] - 2019-12-23
 ### Changed
 - Change flexible layout render values to `client`.